+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/attachments/preview/MediaGalleryPreviewActivity.kt
@@ -627,6 +627,7 @@ public class MediaGalleryPreviewActivity : AppCompatActivity() {
                 handleResult(
                     MediaGalleryPreviewResult(
                         messageId = message.id,
+                        parentMessageId = message.parentId,
                         resultType = MediaGalleryPreviewResultType.SHOW_IN_CHAT
                     )
                 )
@@ -635,6 +636,7 @@ public class MediaGalleryPreviewActivity : AppCompatActivity() {
                 handleResult(
                     MediaGalleryPreviewResult(
                         messageId = message.id,
+                        parentMessageId = message.parentId,
                         resultType = MediaGalleryPreviewResultType.QUOTE
                     )
                 )